refactor: github icon 추가 및 header 구조 변경
