+++ src/components/common/Header/index.tsx
@@ -4,6 +4,8 @@ import Navigation from '../Navigation';
 function Header() {
   return (
     <Styled.Root>
+      <Styled.Logo to="/">seongminn.dev</Styled.Logo>
+
       <Navigation />
     </Styled.Root>
   );
